configurator: Re-organize module layout
Code in `fdt.rs` do not contain architecture specific code as of
`aarch64`, it's only used by `aarch64`, move it one level up to allow
reusing on riscv64 platform.
