--- src/configurator/aarch64/mod.rs
@@ -1,7 +1,0 @@
-// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-//
-// SPDX-License-Identifier: Apache-2.0 AND BSD-3-Clause
-
-//! Traits and structs for configuring and loading boot parameters on `aarch64`.
-
-pub mod fdt;
