+++ src/configurator/mod.rs
@@ -28,9 +28,8 @@ mod x86_64;
 pub use x86_64::*;
 
 #[cfg(target_arch = "aarch64")]
-mod aarch64;
-#[cfg(target_arch = "aarch64")]
-pub use aarch64::*;
+pub mod fdt;
+
 use std::cmp::max;
 use std::mem::size_of;
